Fetch RPC_SPEC_PATH env variable when present
According to the README:
>By default, the type matcher for Gruf controllers matches in `/spec/rpc`.
You can customize this by overriding the value of `ENV['RPC_SPEC_PATH']`
before the require of 'gruf/rspec' in `spec_helper.rb`

This change fetches the value of `RPC_SPEC_PATH` if present and then
falls back to whatever is configured. If nothing is configured, the
default value is used from when `.reset` was previously called.